feat: add SellerPayout entity
