+++ src/controller/payout-request-controller.ts
@@ -24,8 +24,8 @@ import { RequestWithToken } from '../middleware/token-middleware';
 import { parseRequestPagination } from '../helpers/pagination';
 import PayoutRequestService, { parseGetPayoutRequestsFilters } from '../service/payout-request-service';
 import { PayoutRequestStatusRequest } from './request/payout-request-status-request';
-import PayoutRequest from '../entity/transactions/payout-request';
-import { PayoutRequestState } from '../entity/transactions/payout-request-status';
+import PayoutRequest from '../entity/transactions/payout/payout-request';
+import { PayoutRequestState } from '../entity/transactions/payout/payout-request-status';
 import PayoutRequestRequest from './request/payout-request-request';
 import User from '../entity/user/user';
 import BalanceService from '../service/balance-service';
