+++ src/database/database.ts
@@ -52,8 +52,8 @@ import ProductImage from '../entity/file/product-image';
 import BannerImage from '../entity/file/banner-image';
 import StripeDeposit from '../entity/stripe/stripe-deposit';
 import StripePaymentIntentStatus from '../entity/stripe/stripe-payment-intent-status';
-import PayoutRequest from '../entity/transactions/payout-request';
-import PayoutRequestStatus from '../entity/transactions/payout-request-status';
+import PayoutRequest from '../entity/transactions/payout/payout-request';
+import PayoutRequestStatus from '../entity/transactions/payout/payout-request-status';
 import LDAPAuthenticator from '../entity/authenticator/ldap-authenticator';
 import AssignedRole from '../entity/rbac/assigned-role';
 import VatGroup from '../entity/vat-group';
